requirements.txt
Flask==2.2.2
prometheus-client==0.15.0
python-dateutil==2.8.2
requests==2.28.1
urllib3==1.26.12
PyYAML==6.0
MIDIUtil==1.2.1